-> Start fixing spacing inconsistencies occurred due to native header porting -> Fixed a size bug in `SetFloatPropertyValue` in the `AudioObject` class. -> Made all the HAL tests integration tests.

+++ src/NAudio.MacOS/MacOS/CoreAudio/AudioObject.cs
@@ -308,7 +308,7 @@ public abstract class AudioObject :
     {
         ThrowOnStatusError(
             address,
-            NativeMethods.AudioObjectSetPropertyData(objectId, address, sizeof(uint), new(&value))
+            NativeMethods.AudioObjectSetPropertyData(objectId, address, sizeof(float), new(&value))
         );
     }
